+++ ops/hpack.nix
@@ -1,0 +1,103 @@
+{ config, lib, ... }:
+with builtins;
+with lib;
+let
+
+  mergeAttr = a: b:
+  if isAttrs a
+  then merge a b
+  else if isList a
+  then a ++ b
+  else b;
+
+  merge = l: r:
+  let
+    f = name:
+    if hasAttr name l && hasAttr name r
+    then mergeAttr l.${name} r.${name}
+    else l.${name} or r.${name};
+  in genAttrs (concatMap attrNames [l r]) f;
+
+  paths = name: {
+    when = {
+      condition = false;
+      generated-other-modules = ["Paths_${replaceStrings ["-"] ["_"] name}"];
+    };
+  };
+
+  meta = {
+    license = "BSD-2-Clause-Patent";
+    license-file = "LICENSE";
+    author = "Torsten Schmits";
+    maintainer = "hackage@tryp.io";
+    copyright = "2022 Torsten Schmits";
+    category = "Error";
+    build-type = "Simple";
+  };
+
+  base = {
+    name = "base";
+    version = ">= 4.12 && < 5";
+    mixin = "hiding (Prelude)";
+  };
+
+  options.ghc-options = [
+    "-Wall"
+    "-Wredundant-constraints"
+    "-Wincomplete-uni-patterns"
+    "-Wmissing-deriving-strategies"
+    "-Widentities"
+    "-Wunused-packages"
+  ];
+
+  dependencies = [base];
+
+  project = name: doc: merge (meta // { library = paths name; } // options) {
+    inherit name;
+    description = "See https://hackage.haskell.org/package/${name}/docs/${doc}.html";
+    library = {
+      source-dirs = "lib";
+      inherit dependencies;
+    };
+    default-extensions = config.ghci.extensions;
+  };
+
+  exe = name: dir: merge (paths name // {
+    main = "Main.hs";
+    source-dirs = dir;
+    inherit dependencies;
+    ghc-options = [
+      "-threaded"
+      "-rtsopts"
+      "-with-rtsopts=-N"
+    ];
+  });
+
+in {
+
+  polysemy-resume = merge (project "polysemy-resume" "Polysemy-Resume") {
+    synopsis = "Polysemy error tracking";
+    library.dependencies = [
+      base
+      "polysemy >= 1.6"
+      "incipit-core >= 0.2"
+      "transformers"
+    ];
+    tests.polysemy-resume-unit = exe "polysemy-resume" "test" {
+      dependencies = [
+        base
+        "hedgehog"
+        "incipit-core"
+        "polysemy"
+        "polysemy-plugin"
+        "polysemy-resume"
+        "polysemy-test"
+        "stm"
+        "tasty"
+        "tasty-hedgehog"
+        "text"
+      ];
+    };
+  };
+
+}